Lastest round of changes that have the namespace figured out and PSR4 compliance.
