+++ interface/weno/admin.php
@@ -10,11 +10,13 @@
  */
 
 use OpenEMR\Core\Header;
+use Weno\repository\adminProperties;
+
 require_once('../globals.php');
-require_once('transmitDataClass.php');
-require_once("adminClass.php");
 require_once("$srcdir/options.inc.php");
 
+
+/** @var TYPE_NAME $tables */
 $tables   = new adminProperties();
 
 ?>
@@ -27,6 +29,10 @@ $tables   = new adminProperties();
 <body class="body_top">
 <div class="container">
 <?php
+/** check to make sure only administrators access this page. */
+if (!acl_check('admin', $GLOBALS['authUser'])) {
+    die("You are not authorized!");
+}
 
 if($GLOBALS['weno_rx_enable'] != 1){
     print xlt("You must activate Weno first! Go to Admnistration, Globals, Connectors");
@@ -39,7 +45,8 @@ if($GLOBALS['weno_rx_enable'] != 1){
 
    $drugData = $tables->drugTableInfo();
 if(!$drugData['ndc']){
-    echo "<a href='drugPaidInsert.php' class='btn btn-default'>".xlt("Import Formularies")."</a> <br>".xlt("Be patient this may take a while");
+    echo "<a href='drugPaidInsert.php' class='btn btn-default'>".xlt("Import Formularies").
+         "</a> <br>".xlt("Be patient this may take a while");
 } else {
     print xlt("Formularies inserted into table")."<br>";
 
@@ -52,18 +59,24 @@ if(!$drugData['ndc']){
 
 <form method="post" action="import_pharmacies.php" >
     <div class="col-lg-2">
-    <?php echo generate_form_field(array('data_type'=>$GLOBALS['state_data_type'],'list_id'=>$GLOBALS['state_list'], 'field_id'=>'state')); ?>
+    <?php
+    $currvalue = "";
+    $frow = array('data_type'=>$GLOBALS['state_data_type'],'list_id'=>$GLOBALS['state_list'],'field_id'=>'state');
+    echo generate_form_field($frow, $currvalue); ?>
     </div>
     
-
-    <button type="submit" class="btn btn-default btn-save" value= ><?php echo xlt("Import Pharmacies"); ?> </button>
+    <button  class="btn btn-default btn-save" value= ><?php echo xlt("Import Pharmacies"); ?> </button>
 
     <br>
-<p><?php echo xlt("Be patient, this can take a while."); ?><br> </p>
+<p><?php echo xlt("Be patient, this can take a while."); ?> </p>
 </form>
-<br><br>
+<br>
+    <br>
 
-<?php  if(!empty($finish)){echo $finish . xlt("with import");} ?>
+<?php  if (!empty($finish)) {
+    echo $finish . xlt("with import");
+}
+?>
 
 
 </div>
