+++ interface/weno/jsonScript.php
@@ -9,9 +9,9 @@
  * @license https://github.com/openemr/openemr/blob/master/LICENSE GNU General Public License 3
  */
 
-include_once("../globals.php");
-include_once($srcdir."/patient.inc");
-include_once("transmitDataClass.php");
+require_once("../globals.php");
+require_once($srcdir."/patient.inc");
+require_once("transmitDataClass.php");
 
 $date = date("Y-m-d");
 $pid = $GLOBALS['pid'];
